microservices/inventory-service/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
python-jose
